requirements.txt
openai
langchain
faiss-cpu
fastapi
tiktoken
python-dotenv
pillow
langchain-core
streamlit
langchain-experimental
tabulate
pypdf
gradio